Deny unknown fields in butido's configuration
I already implemented this in 00636cf but forgot to also deny unknown
fields in all sub-structures (recursively). This fixes that so now it
shouldn't be possible to have any unknown fields in butido's
configuration without getting an error message (I already had to remove
`speed` in an example configuration to fix the tests).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>