Show relative paths instead of warnings for missing releases
When listing releases from the DB, this will replace the warnings for
paths to release files that aren't available locally with relative paths
and a note that they aren't available locally.

The warnings were present since the initial implementation (19cc89d) but
don't make much sense nowadays as butido can be used on multiple systems
(clients) so it can be expected that not all releases will be available
locally (plus there could be a cleanup/GC policy in place).
This change drops those warnings to improve/cleanup the output by
sticking to a table, which also ensures that all relevant information is
there (the warnings didn't contain the release dates).
We could introduce a new column to indicate if a given release is
available locally but deemed it nicer to integrate that information into
the "Path" column. This has the advantage that we don't need more space
and another advantage is that the paths will be less confusing as the
prefix comes from the local configuration file (`releases_root` setting
vs. data from the DB) so the actual path on the system where the release
is present might be different. Displaying the relative paths (vs. an
empty path or "-") has the advantage that the output still contains all
relevant output from the DB (store and file names).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>