Apply the usize parser to all LIMIT parameters for better error messages
This ensures that the errors will be emitted at the right time/place so
that it is clear which CLI parameter caused the parsing error.

Old output:
```
$ butido db releases -L a
Error: invalid digit found in string
```

New output:
```
$ butido db releases -L a
error: invalid value 'a' for '--limit <LIMIT>': invalid digit found in string

For more information, try '--help'.
```

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>