Properly parse the `--limit` CLI parameter from a string to usize
We did the parsing twice as the unnecessary `parse_usize` function did
convert the string to usize but then back to a string again (instead of
returning the usize). This switches to Clap's generic value parser which
should be equivalent with the old custom function without the
`map(|_| s.to_owned())` (returning `Result<usize, String>`).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>