Update all dependencies (Cargo.lock) and bump the MSRV (1.76 -> 1.78)
This is simply the result of running `cargo update` without touching
Cargo.toml (required since we currently only use dependabot to update
direct dependencies and not indirect/transitive dependencies and useful
to avoid a lot of dependabot PRs for minor/patch updates that can be
bundled, like in this commit).

The crate `diesel_derives v2.2.0` requires `rustc 1.78.0` or newer so
let's bump the MSRV accordingly.

Bumping the MSRV also removes the last usage of the `rustversion` crate
but I'll keep it around for now as it can become relevant again with
every new Rust version (so it seems best to just keep it around instead
of having to remove and re-add it "regularly").

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>