Fix the ASCII table rendering for data with colored strings
The colored output works via control characters that are part of the
strings that are put in the columns of the table. This resulted in an
incorrect column length computation as the `color_codes` feature is
required to correctly support terminal color codes.
I decided to additionally enable `wide_characters` support in case we
ever start using (or are already using?) wide characters [0] in the
data/output (e.g., emojis). That feature only requires an additional
dependency on `unicode-width` that is already in `Cargo.lock`.

Colors are, e.g., used in the output of "butido db submit $SUBMIT".

[0]: https://en.wikipedia.org/wiki/Wide_character

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>