Turn the warnings for duplicate container images names into errors
There must not be any name collisions for the algorithms to expand and
shorten image names to work correctly so we shouldn't only emit warnings
given that these collisions are trivial to fix by simply removing the
collisions from the provided configuration.

This also improves the output and warns about collisions between a full
and a short name again (I initially skipped this when switching to two
separate hash maps in 260b4d7).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>