Fix the error propagation of the "source download" command
The additional source verification in 2f8034b introduced a regression by
returning the source verification errors before the errors of the
download operation (the download errors therefore get omitted and users
get misleading errors that the sources are missing and that the
verification did therefore fail).

This restores the proper error messages and avoids unnecessary source
verification by returning errors of the download operation before
triggering the verification.

We also don't need the debug output as we already return the errors and
I've put `Ok(())` at the end to make it a bit more readable.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>