Support PackageVersionConstraint specification without a "constraint"
This is mainly done for consistency.
There are currently two ways to specify package/dependency versions:
- `PackageVersion`: An exact version.
- `PackageVersionConstraint`: A "constraint" character followed by a
  version. Only `=` is currently supported as "constraint" character
  though, so the version must be an exact match as well.

This means that `PackageVersionConstraint` currently only implements the
exact same version matching logic as `PackageVersion` but additionally
requires the `=` character in front of the version string.
This makes that "constraint" character optional so that it doesn't
trigger a parsing error when, e.g., a CLI argument lacks it (it was a
bit annoying that some subcommands use `PackageVersionConstraint` and
therefore required the `=` in front of the version).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>