Only support a single version for the `env-of` command
Using a `PackageVersionConstraint` can result in multiple versions of
the package matching, in which case butido will output the environment
variables of all packages/versions but without any indication which
variable belongs to which version (optionally with some variable merging
logic). We could support this but for now it seems best to explicitly
only support a single version.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>