Fix the help/usage output and parsing of the "release new" subcommand
The `--all` flag didn't work properly as it expected a value.
The old usage (from the error/help output) looks like this:
> butido release new [OPTIONS] --to <RELEASE_STORE_NAME> <PKG|--all <all-packages>> <SUBMIT> [VERSION]

The position/index of `PKG` and `SUBMIT` are wrong. The `PKG` argument
should come after `SUBMIT` but it likely appears before `SUBMIT` in the
usage output as it's in a `ArgGroup` with the `--all` flag and
flags/options come before arguments with values.
I decided to simply swap the indices of `PKG` and `SUBMIT` as it's more
important that the usage matches the parsing (-> no unexpected errors)
than that it's quicker / more ergonomic to repeat the command with the
same submit but a different package.

The new usage is now consistent with the parsing:
> butido release new [OPTIONS] --to <RELEASE_STORE_NAME> <PKG|--all> <SUBMIT> [VERSION]

It isn't ideal though that `SUBMIT` is now between `PKG` and `VERSION`!
(But it currently seems to be the only combination that works right...)

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>