Properly fix the "release new" subcommand by dropping `--all`
I noticed that the approach from the previous commit sometimes(?) still
resulted in some issues and it looks like it's a bad idea to mix
options/flags and positional arguments/values in a `ArgGroup`.

The `all-packages` argument isn't even used in the code (it's
sufficient that `package` is not set) so it seems way batter to simply
drop the `--all` flag/option and keep `package_name` optional (I've
extended the description a bit to properly document the behaviour).

I've also restored the original indices of the positional
values/arguments to make the usage more ergonomic again.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>