Fix making the package name/version arguments required ("release rm")
The code calls `unwrap()` on both arguments and therefore panics if any
of them are not set (`None`). This fixes the specification of both
arguments so that the usage is correct and that Clap will emit an error
instead of butido panicking/crashing.
It also makes sense that both of these arguments are required to avoid
accidental deletion of multiple versions up to nuking an entire release
store (there's a confirmation prompt but accidentally hitting `y` is
enough to confirm it).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>