Let the `download_manually` source property default to `false`
This is an optional setting that should only have to be set for sources
that have to be downloaded manually. The example in
`examples/packages/repo/pkg.toml` isn't a proper default as it only
applies to the source "src" (packages could have multiples sources or
choose a different name).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>