Display the artifact file suffix/type in the "db jobs" output
This adds a new column ("Type") to the table that "butido db jobs"
outputs. It works under the assumption that the artifact file names
indicate the type of artifact via a file name extension [0] (i.e., using
the format "{name}.{type}").
The fallback values are "-" if no artifact is available (build failures,
aborted jobs, etc. - that's also why we're using LEFT JOIN) or "?" (for
unknown) if no file name extension was detected.

This resolves #391.

[0]: https://en.wikipedia.org/wiki/Filename_extension

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>