Update all dependencies (Cargo.lock)
This is simply the result of running `cargo update` without touching
Cargo.toml (required since we currently only use dependabot to update
direct dependencies and not indirect/transitive dependencies and useful
to avoid a lot of dependabot PRs for minor/patch updates that can be
bundled, like in this commit).

Because 19 new(/updated?) dependencies use the `Unicode-3.0` license, we
have to add it to the list of allowed licenses in `deny.toml`.
This should be fine given that we already allow the `Unicode-DFS-2016`
license and both licences are OSI approved [0][1].
On the SPDX license list [0] both Unicode licenses aren't listed as "FSF
free/libre" but this is likely only because those versions have never
been checked by FSF. The FSF lists the "Unicode, Inc. License Agreement
for Data Files and Software (#Unicode)" license under "Free licenses,
compatible with the GNU GPL" [2] and comments that "It is a lax
permissive license, compatible with all versions of the GPL." but
the license link [3] redirects to the `Unicode-DFS-2012` version.

[0]: https://spdx.org/licenses/
[1]: https://opensource.org/license/unicode-license-v3
[2]: https://www.gnu.org/licenses/license-list.html#Unicode
[3]: https://directory.fsf.org/wiki/License:Unicode

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>