Let cargo-deny ignore that the instant crate is unmaintained
This isn't a security issue yet, the `indicatif` crate should get rid
of it eventually, and this isn't really actionable for us so let's
ignore that informative advisory so that CI remains green (it's an
optional CI check but the failure could mask real vulnerabilities and
does also mark the CI run with a red cross on the GitHub UI).

Thank you <michael.weiss@eviden.com> for the nice copypasta commit
message from the last deny unmaintained advisory.

Signed-off-by: Nico Steinle <nico.steinle@eviden.com>