Update the `config` crate from 0.13.4 to 0.15.4
We had to revert the update to version 0.14 in 4c6f027 due to a
breaking change that made the `config` crate case-insensitive regarding
configuration keys/settings.
The impact of that upstream change/"regression" was tracked in [0],
the change finally got reverted in [1], and version 0.15.0 was released
with the original (case-sensitive) behaviour.

[0]: https://github.com/mehcode/config-rs/issues/531
[1]: https://github.com/rust-cli/config-rs/pull/613

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>