Replace the `daggy` crate with `petgraph`
The `daggy` crate doesn't seem maintained anymore as the last commit was
over three years ago [0] and there isn't really any activity since then.

Daggy is already based on top of `petgraph` and described as follows:
> The most prominent type is Dag - a wrapper around petgraph’s Graph
> data structure, exposing a refined API targeted towards directed
> acyclic graph related functionality.

This means that we can switch directly to `petgraph` without too much
effort. We'll loose some of the refined API, especially walking graphs
via iterators, but it doesn't really matter too much as the `Acyclic`
type, "a wrapper around graph types that enforces an acyclicity
invariant", works well enough (just a bit less "refined").

We also already used `petgraph` directly for the `tree-of` command.
This direct dependency on `petgraph` became the trigger for dropping the
dependency on `daggy` since `petgraph` yanked the release of version
`0.6.6` with the new release of version `0.7.0` [1] and the resulting CI
failure for the `petgraph` update [2] made me take a closer look at the
situation (we don't necessarily have to drop `daggy` just yet but it
seems for the best given it's unclear future and the duplicated
`petgraph` dependency that causes incompatibilities / build failures).

[0]: https://github.com/mitchmindtree/daggy
[1]: https://github.com/petgraph/petgraph/issues/712
[2]: https://github.com/science-computing/butido/pull/446

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>