Fix the CLI documentation regarding package version constraints
We might want to add an abstraction for those arguments in the future to
make this easier but at least the arguments are now all named
`package_version_constraint` with the argument value name
`VERSION_CONSTRAINT` (e.g., in the help output), and a help text /
description that makes clear that this is a version constraint and not a
single, exact version.

The documentation of the package version arguments for the
`source {verify,url,of}` subcommands didn't match the code
(a `PackageVersionConstraint` is used in the code).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>