Make `semver::Op::Exact` the default for `PackageVersionConstraint`
This seems like a much better and safer default than
`semver::On::Caret`, which I chose in fe0df00.

It's more intuitive and it's also the best way to support both our
strict implementation for `=` (full version matches only) as well as
`semver::Op::Exact` (partial version matches are supported).

I don't remember why I previously picked `^` (Caret / "compatible"
updates) as the default - probably just because it's the default that
the `semver` crate uses [0] (since the crate is for Cargo's flavor of
SemVer [1]) and I didn't spend much thought on it.

[0]: https://github.com/dtolnay/semver/blob/1.0.24/src/parse.rs#L158
[1]: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#caret-requirements

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>