Fix the fallback for parsing `PackageVersion` into `semver::Version`
I later noticed that the previous iteration didn't work as "recursive"
capture groups aren't supported (so the nested numbered capture group
for the version number(s) in the outer non-capturing group that may
match zero or more times doesn't result in more than one capturing group
and only the last match will be returned for the capturing group).
This new logic fixes that but it's still far from ideal...

I also noticed that butido doesn't seem to do any verification of the
version numbers using `PackageVersion::parser()` so far - at least in
the code branches that I tested so we definitely should change that.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>