Fix the error handling for parsing `PackageVersionConstraint`s without =
We currently support two kinds of `PackageVersionConstraint`:
1) Exact version matches (`=$VERSION`): Those are implemented using
   string comparison (`==`).
2) Exact but partial version matches (`$VERSION`): Those are implemented
   using the `semver` crate and `semver::Op::Exact` (`=$VERSION`).

In the second case the version number/string must be a valid
`semver::VersionReq` which is more strict than our butido versions
(especially since we have a parser for `PackageVersion` but it isn't
used to validate the versions yet).
We didn't check this in `PackageVersionConstraint::parser()` before so
butido would panic in `PackageVersionConstraint::matches()` instead of
returning an appropriate error message if the provided
`PackageVersionConstraint` isn't "valid".
This fixes that by attempting to parse the version using the `semver`
crate in the `parser()` function as well (we need it to parse it as
`VersionReq` since `Version` must be a full version (including the minor
and patch versions) - I therefore decide to deduplicate
`default_constraint` using a private function).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>