Switch the package dependency resolving to normal `PackageVersion`s
The current logic was never capable of properly handling
`PackageVersionConstraint` but this wasn't an issue so far since the
only supported constraint was an exact version match using `=`.

I decided to temporarily switch this code to just `PackageVersion`
before properly implementing this to avoid potential "regressions".
(The current code works fine with `PackageVersionConstraint` but the
logic doesn't make sense as all matching versions would be build
instead of only the newest one - this isn't an issue as long as the
`pkg.toml` files aren't changed to use versions that aren't prefixed
using `=` but let's avoid this potential issue for now.)

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>