Switch to the `anyhow::bail!` macro where possible
This does basically apply 354db2a to the whole codebase for consistency.
The `anyhow::bail!` is basically syntactic sugar that is a bit more
compact (tbh I'm a bit split on it as it also hides the `return` a bit
but it's name is expressive enough and since we already make use of it
we might as well also use it consistently everywhere where possible).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>