Update all dependencies (Cargo.lock) and bump the MSRV (1.80 -> 1.85)
This is simply the result of running `cargo update` without touching
Cargo.toml (required since we currently only use dependabot to update
direct dependencies and not indirect/transitive dependencies and useful
to avoid a lot of dependabot PRs for minor/patch updates that can be
bundled, like in this commit).

The MSRV bump to 1.85 (currently the latest stable version) is required
since at least the dependency `ascii_table-4.0.6` already switched to
Rust edition 2024 which was stabilized in the 1.85 release [0].
It's a bit early for the bump but it's fine for us since we're not a
library, our dependencies are already requiring it, and I already
considered switching to the new Rust edition anyway.

[0]: https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#edition-2024

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>