Update the Rust edition from 2021 to 2024
See "The Rust Edition Guide" [0] for details regarding the changes in
Rust 2024. I started with `cargo fix --edition` [1] but ended up
reverting the "unnecessary" `if let` to `match` conversions [2] and had
to preform a manual change for the tail expression temporary scope [3]
(in the `container_stats()` function).
Other relevant changes are [4] (we could revert the change in `path.rs`)
and [5].

I also manually updated the edition in `rustfmt.toml` and set the new
`style_edition` [6] to `2021` to split the formatting changes into a
dedicated commit (will be done later).

[0]: https://doc.rust-lang.org/edition-guide/rust-2024/index.html
[1]: https://doc.rust-lang.org/edition-guide/editions/transitioning-an-existing-project-to-a-new-edition.html
[2]: https://doc.rust-lang.org/edition-guide/rust-2024/temporary-if-let-scope.html
[3]: https://doc.rust-lang.org/edition-guide/rust-2024/temporary-tail-expr-scope.html
[4]: https://doc.rust-lang.org/edition-guide/rust-2024/rpit-lifetime-capture.html
[5]: https://doc.rust-lang.org/edition-guide/rust-2024/match-ergonomics.html
[6]: https://rust-lang.github.io/rustfmt/?version=v1.8.0&search=edition#style_edition

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>