Inline variables in format strings
This is done via Clippy by enabling the pedantic lint
`uninlined_format_args`. The lint was moved from `pedantic` to `style`
in Rust 1.88 but also back to `pedantic` in Rust 1.89 [1]. We can apply
the recommended changes so let's do it to make CI happy (lints with the
MSRV, i.e., Rust 1.88).

[0]: https://rust-lang.github.io/rust-clippy/master/index.html#uninlined_format_args
[1]: https://github.com/rust-lang/rust-clippy/blob/432dad4dbe28801adeb129f25981283c7ff18aa4/CHANGELOG.md

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>